rust: Search all files in install.yml from outputdir first
For `rust` packages, all the files will be searched inside of the output
directory (usually `target`) and, if they don't exist, inside the project
directory. This allows rinstall to correctly get files generated by
`build.rs` without adding too much clutter in the `install.yml`.
To force rinstall to get a file in the project directory, use the
`$PROJECTDIR` placeholder.

Example `install.yml` that will be improved by this commit:

https://github.com/kanidm/kanidm/blob/683e0ca8e3a7325434331de5773b21e5b51b4cea/install.yml